mkosi: Adjust kernel modules in ramdisk
It's been observed that including all the kernel modules in the ramdisk
makes it too big for the default EFI System Partition size. At the same
time, mkosi has introduced the "default" meta-kernel-module, so we can
use that to tackle most of the modules that needs to be included.
Unfortunately a few dependencies are missing for btrfs (used by Fedora),
so those needs to be spelled out - in addition to this, we need the
Qualcomm storage drivers.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>