mkosi: Switch to Unified Kernel Image (UKI)
Prior issues with generating Unified Kernel Images has been resolved and
ukify has the benefit of automatically unpacking the kernel if the
installed one is compressed. This means that switching to UKI removes
the need for CONFIG_EFI_ZBOOT, and hence the upstream arm64
defconfig works out of the box.

systemd-boot-efi is added in Debian, to get the necessary systemd-boot
stub.

Retain the comment in the README, for those who tinker with the
configuration, but move it down and make it less significant.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>