mkosi: Increase deferred_probe_timeout for all targets
The deferred_probe_timeout was explicitly added to the QCS6490-based
targets after observing issues with the kernel not booting fast enough.

But as we no longer bundle all the kernel modules in the ramdisk, some
of the clock controllers (in particular gpucc) aren't available until
the rootfs is mounted and module loading has been attempted again.

In general targets gets past the ramdisk in the alotted time, but it
seems like a good idea to extend this universally, to avoid those cases
where it doesn't.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>